pep and some fixes
peps: https://pylint.readthedocs.io/en/latest/user_guide/messages/messages_overview.html
discord is an outdated library, use nextcord or other libraries
