+++ .pylintrc
@@ -1,0 +1,2 @@
+[MESSAGES CONTROL]
+disable = missing-docstring
